+++ src/main/java/util/NumberGenerator.java
@@ -1,4 +1,4 @@
-package model;
+package util;
 
 public interface NumberGenerator {
     int generateNumber();
