+++ src/main/java/util/RandomNumberGenerator.java
@@ -1,4 +1,4 @@
-package model;
+package util;
 
 public class RandomNumberGenerator implements NumberGenerator {
     private static final int BOUND = 10;
